Added "expect" feature for services.
Initially added to deal with RPC, this feature makes possible to filter the services, locally or globally.

Documentation:
**************
A function to execute before the success. If returns true, the "success" callback will be triggered. Else, the "error" callback will be triggered. This value can be set as well from the instance settings or the global settings.
This function takes as arguments the data returned by the service, the input object and the service configuration.