Fixed #11: Global ref to the instance definition
Added the method "configuration" in every global instance that returns a deep clone of the initially given configuration hash. Also, calling this method with a string as parameter will return the corresponding part of the definition ("properties", "hacks", "name", ...).
This object is cloned at every call, to make sure it is not possible to modify the instance definition through this method.