Fixed #20: Added an "update" method in scopes
Since there are now methods in scopes that work asynchronously ("request", "dispatchEvent"), I added an "update" method to update one or several properties.
It has the exact same effect that updating the property directly inside the scope - and both ways to update are working.
The reason I added a new way to update properties is because being able to write properties in the scope but unable to read them is not really consistant.
I will encourage people to use this fresh new "update" method from now.