Fixed #30: "includes" flag in custom structures
First, it is now possible to give to domino.struct.check as third argument an object containing flags. Currently, the only recognized flag is "includes", and it allows non-specified keys to be used additionally to the ones required.

Also, any flag given to a custom structure when added is now also given to domino.struct.check, so using this new "includes" flag in a custom structure works.