Fixed #26: descriptions management + help() method
The "description" attribute is now recognized for shortcuts, hacks, properties and services. It is still optional, and has to be a string if specified.

The global instance also has from now a "help()" method, to see these properties. See the method documentation or the unit tests to learn exactly how it works.