Switched from Makefile to Grunt.js
Added Gruntfile.js and package.json to use Grunt tasks instead of a custom Makefile.

The unit tests are still startable from the test/unit.html in the browser, but now also can be executed as a Grunt task:
  > grunt qunit

Also, the Closure compiler is no more used, and is replaced by Uglify:
  > grunt uglify

And the Google Closure lint still works, with the command:
  > grunt closureLint

Finally, the command `grunt` will successively execute the tasks `closureLint`, `qunit` and `uglify`.