Fixed #42: Improved "before" feature in services
First, I added the "beforeSend" callback to the domino Ajax utils. It works exactly as in jQuery.

Also, services "before" now take a reference to the XHR object as second argument.

Domino generates a "beforeSend" callback from the "before" function to make sure it will have an access to the XHR object BEFORE the call is sent.

Also: unit tested.