Fixed #44: Added log, warn and die in modules
From now, when a domino instance adds a module, it also binds its `log`, `warn` and `die` methods to the module.

The logs will be prefixed by `[module]`, or `[module "id"]` if the module has the id "id".

Also added `domino.utils.partial`, that works exactly as `_.partial`. Unit tested.