Added events triggerer
It works with the "on", "off" and "trigger" methods, but the code is exactly the same as in the original domino.
Only modifications: No more getEvent method, no more split on spaces. Any string can now be an event name.
Also ported the related unit tests that are still relevant.