Started to draft modules management
The modules constructor is domino.module.js. It is slightly better than in domino v1, with a better wording:
 - A reference to the controller is given at this.constructor in the module, instead of sending that as an argument.
 - The constructor listens to every events sent from the module, wether they are referenced in the controller or not.

The main issue that is still a problem is that I wanted to bind events in the module through this.controller.on(), but this makes unbinding these events when the module is killed very difficult.